expand boundary element module to multiple files
